Brings wasabipesto/nice up to a066ab3 (origin/main). #87 came in on the
previous merge; this one is #88 plus the version/changelog commits.

Conflict: #88 edited GPU_LSD_K and the MSD-floor machinery in place in
client_process_gpu.rs, but this branch had already moved both into
gpu_niceonly.rs so the Vulkan backend could share them. Resolved by
keeping the shared module and carrying #88's semantic change to it —
GPU_LSD_K 2 -> 3. The duplicated AdaptiveFloor block upstream re-added
was dropped; it is identical to the one already in gpu_niceonly.rs.

The k=3 table needed one real fix on the Vulkan side. The niceonly
shader reduces a range offset with a Horner loop, acc = (acc << c | chunk)
% M, which requires M <= 2^(32-c); c was a fixed 8, sound only because
M = (b-1)*b^k stayed under 2^24 at k=2. At k=3 every modulus gains a
factor of b: base 65 hits 17 576 000 and base 128 hits 266 338 304, so
NiceonlyConfig::new would have refused every base >= 65 — base 80, which
is in production here, among them. stride_chunk_bits now picks c per base
(8 while M <= 2^24, else 4, bound 2^28), and the emitted shader and the
host mirror derive it from the same modulus. It costs nothing measurable:
this reduction runs once per range descriptor, not per candidate.

Also adapted this branch's stride-table test mirrors to the u32 residue
and gap types, the same way #88 did for the CUDA ones, and added base 80
to the device parity list — every base below 65 keeps the byte chunk, so
nothing else here executes the 4-bit path on hardware.

Verified: 120 unit tests plus all 6 device tests pass on the AMD iGPU;
niceonly and detailed both match the CPU at bases 10/12/25/40/45/62/80.

+++ common/src/vulkan/mod.rs
@@ -1372,7 +1372,7 @@ mod tests {
                 if codegen::mirror_prefilter(n, &kernel, &pre).0 {
                     want.push(n);
                 }
-                n += table.gap_table[idx];
+                n += u128::from(table.gap_table[idx]);
                 idx = (idx + 1) % table.gap_table.len();
             }
